+++ packages/test/jest.config.ts
@@ -8,14 +8,12 @@ module.exports = {
         diagnostics: {
           exclude: ['**'],
         },
+        tsconfig: './tsconfig.jest.json',
       },
     ],
-    '^.+\\.(js|jsx)$': 'babel-jest',
+    '^.+\\.[t|j]sx?$': 'babel-jest',
   },
-  setupFilesAfterEnv: [
-    '@testing-library/jest-dom/extend-expect',
-    '<rootDir>/__mocks__/consoleMock.ts',
-  ],
+  setupFilesAfterEnv: ['<rootDir>/__mocks__/consoleMock.ts'],
   moduleNameMapper: {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/__mocks__/fileMock.ts',
@@ -23,4 +21,5 @@ module.exports = {
   },
   transformIgnorePatterns: ['/node_modules/(?!(@testing-library|@baseapp-frontend)/)'],
   verbose: true,
+  moduleDirectories: ['node_modules', '<rootDir>'],
 }