+++ packages/test/providers/ComponentWithProviders/index.tsx
@@ -1,9 +1,9 @@
-import React, { FC, PropsWithChildren, ReactNode } from 'react'
+import { FC, PropsWithChildren } from 'react'
 
 import { ThemeProvider, theme } from '@baseapp-frontend/design-system-mui'
 import { ReactQueryProvider } from '@baseapp-frontend/provider'
 
-const ComponentWithProviders: FC<PropsWithChildren<ReactNode>> = ({ children }) => (
+const ComponentWithProviders: FC<PropsWithChildren> = ({ children }) => (
   <ReactQueryProvider>
     <ThemeProvider theme={theme} cacheKey="test-key">
       {children}