+++ packages/test/providers/ComponentWithRootLayout/index.tsx
@@ -1,0 +1,9 @@
+import { FC, PropsWithChildren } from 'react'
+
+const ComponentWithRootLayout: FC<PropsWithChildren> = ({ children }) => (
+  <html lang="en">
+    <body>{children}</body>
+  </html>
+)
+
+export default ComponentWithRootLayout